# feat(java): native streaming mode deserialization