--- a/java/fury-core/src/main/java/org/apache/fury/builder/BaseObjectCodecBuilder.java
+++ b/java/fury-core/src/main/java/org/apache/fury/builder/BaseObjectCodecBuilder.java
@@ -93,6 +93,7 @@ import org.apache.fury.serializer.collection.AbstractMapSerializer;
 import org.apache.fury.serializer.collection.CollectionFlags;
 import org.apache.fury.type.TypeUtils;
 import org.apache.fury.util.GraalvmSupport;
+import org.apache.fury.util.Platform;
 import org.apache.fury.util.Preconditions;
 import org.apache.fury.util.ReflectionUtils;
 import org.apache.fury.util.StringUtils;
@@ -252,7 +253,7 @@ public abstract class BaseObjectCodecBuilder extends CodecBuilder {
       return "CutPoint{" + "genNewMethod=" + genNewMethod + ", cutPoints=" + cutPoints + '}';
     }
   }
-
+`
   protected void registerJITNotifyCallback() {
     // build encode/decode expr before add constructor to fill up jitCallbackUpdateFields.
     if (!jitCallbackUpdateFields.isEmpty()) {
@@ -274,8 +275,7 @@ public abstract class BaseObjectCodecBuilder extends CodecBuilder {
    * @see CodeGenerator#getClassUniqueId
    */
   protected void addCommonImports() {
-    ctx.addImports(List.class, Map.class, Set.class);
-    ctx.addImports(Fury.class, MemoryBuffer.class, fury.getRefResolver().getClass());
+    ctx.addImports(Fury.class, MemoryBuffer.class, fury.getRefResolver().getClass(), Platform.class);
     ctx.addImports(ClassInfo.class, ClassInfoHolder.class, ClassResolver.class);
     ctx.addImport(Generated.class);
     ctx.addImports(LazyInitBeanSerializer.class, Serializers.EnumSerializer.class);