… java/benchmark/src/main/java/org/apache/fury/benchmark/MemorySuite.java
… java/fury-core/src/main/java/org/apache/fury/builder/CompatibleCodecBuilder.java
… java/fury-core/src/main/java/org/apache/fury/builder/MetaSharedCodecBuilder.java
… java/fury-core/src/main/java/org/apache/fury/builder/ObjectCodecBuilder.java
… java/fury-core/src/main/java/org/apache/fury/builder/ObjectCodecOptimizer.java
… java/fury-core/src/main/java/org/apache/fury/codegen/ExpressionUtils.java
… java/fury-core/src/main/java/org/apache/fury/io/AbstractStreamReader.java
… java/fury-core/src/main/java/org/apache/fury/io/FuryInputStream.java
… java/fury-core/src/main/java/org/apache/fury/io/FuryReadableChannel.java
… java/fury-core/src/main/java/org/apache/fury/io/FuryStreamReader.java
… java/fury-core/src/main/java/org/apache/fury/io/MemoryBufferInputStream.java
… java/fury-core/src/main/java/org/apache/fury/io/MemoryBufferObjectInput.java
… java/fury-core/src/main/java/org/apache/fury/io/MemoryBufferObjectOutput.java
… java/fury-core/src/main/java/org/apache/fury/io/MemoryBufferOutputStream.java
… java/fury-core/src/main/java/org/apache/fury/io/MemoryBufferReadableChannel.java
… java/fury-core/src/main/java/org/apache/fury/io/MemoryBufferWritableChannel.java
… java/fury-core/src/main/java/org/apache/fury/io/MockOutputStream.java
… java/fury-core/src/main/java/org/apache/fury/io/MockWritableChannel.java
… java/fury-core/src/main/java/org/apache/fury/io/StreamReadable.java
… java/fury-core/src/main/java/org/apache/fury/memory/MemoryBuffer.java
… java/fury-core/src/main/java/org/apache/fury/pool/ThreadPoolFury.java
… java/fury-core/src/main/java/org/apache/fury/resolver/ClassResolver.java
… java/fury-core/src/main/java/org/apache/fury/resolver/FieldResolver.java
… java/fury-core/src/main/java/org/apache/fury/resolver/MapRefResolver.java
… java/fury-core/src/main/java/org/apache/fury/serializer/ArraySerializers.java
… java/fury-core/src/main/java/org/apache/fury/serializer/ExternalizableSerializer.java
… java/fury-core/src/main/java/org/apache/fury/serializer/JavaSerializer.java
… java/fury-core/src/main/java/org/apache/fury/serializer/PrimitiveSerializers.java
… java/fury-core/src/main/java/org/apache/fury/serializer/Serializers.java
… java/fury-core/src/main/java/org/apache/fury/serializer/StringSerializer.java