@chaokunyang 2024-04-03
Hi @Munoon, I implemented streaming deserialization for input stream. Maybe it can be taken as an example for java.nio.Channel. The implementation is a littl tricky, maybe there is a better way.