@Munoon 2024-04-03
> Hi @Munoon, I implemented streaming deserialization for input stream. Maybe it can be taken as an example for java.nio.Channel. The implementation is a littl tricky, maybe there is a better way.

Hi, @chaokunyang 
I feel like I'm ready with my PR, but I'm not sure how to send it, since it is built on top of your PR. I think that I will just wait until this PR will be merged. Let me know if you have a better idea.

Also, some methods of my implementation are fully copied from your `FuryInputStream` and some are partially. I think those can be extracted in the abstract class, but I'll leave it as is to make a PR simpler. Maybe it worth making a refactoring in a separate PR.