@chaokunyang 2024-04-03
> ```java
>   public void increaseReaderIndex(int diff) {
>     int readerIdx = readerIndex + diff;
>     if (readerIdx < 0) {
>       throw new IndexOutOfBoundsException(
>           String.format(
>               "readerIndex: %d (expected: 0 <= readerIndex <= size(%d))", readerIdx, size));
>     } else if (readerIdx > size) {
>       int readCount = streamReader.fillBuffer(diff);
>       if (diff > readCount) {
>         throw new IndexOutOfBoundsException(
>                 String.format(
>                         "readerIndex: %d (expected: 0 <= readerIndex <= size(%d))", readerIdx, size));
>       }
>     }
>     this.readerIndex = readerIdx;
>   }
> ```

Nice catch, I fixed it, but fill buffer by `streamReader.fillBuffer(readerIdx - size);`