@chaokunyang 2024-04-03
I think it's important to our users, so I plan to include this PR in our 0.5.0 release. The original streaming is error-prone when using, and not a real streaming implementation.