@chaokunyang 2024-04-04
> This PR introduce 25% performance degeneration, we can''t merge it

I managed to make it faster than version without streaming reading. I'll split those optimization into other PR, then we can merge this PR later.