@chaokunyang 2024-04-06
New readBytesString are too large for inline:
![image](https://github.com/apache/incubator-fury/assets/12445254/e76ccf29-a30f-4eab-8b16-c817d553dd05)

Old version:
![image](https://github.com/apache/incubator-fury/assets/12445254/458406d6-7a14-4de1-bd7c-fd698865ba67)